requirements.txt
boto3
python-dotenv
manim
mutagen
imageio-ffmpeg
sounddevice
numpy
scipy
gTTS
deepgram-sdk
